Code formatting
Just to appease my OCD; I don't like lines becoming so long they wrap on my end.The filter names are also a bit long, but I like them to be easy to understand in what their purpose is.

Also, I use the `[]` syntax for simple numerical arrays and `array()` for associative ones. There is an argument to use the long one and this is what the WP guide says... but I don't care.